code review: move rari config to src/configs
